% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QREM.R
\name{flatQQplot}
\alias{flatQQplot}
\title{A `flat QQ-plot' for a fitted quantile regression model.}
\usage{
flatQQplot(
  x,
  evalx,
  M,
  qns,
  L = 21,
  minRatio = 0,
  maxRatio = 2,
  filename = NULL
)
}
\arguments{
\item{x}{The variable used in the QQ-plot.}

\item{evalx}{The values of the variable x where the quantiles are calculated.}

\item{M}{A matrix obtained from QRdiagnostics, applied to multiple quantiles.}

\item{qns}{The quantiles for which we compute the QQ plot.}

\item{L}{The number of color levels to be used in the heatmap (default=21).}

\item{minRatio}{The smallest ratio between expected and observed quantile (default=0).}

\item{maxRatio}{The largest ratio between expected and observed quantile (default=2).}

\item{filename}{The pdf file to save the plot. Default is NULL (print to the screen.)}
}
\description{
Showing multiple ('flat') QQ plots for different quantiles using a heatmap.
}
\examples{
\donttest{
data(simdf)
L <- 20
qns <- seq(0.1,0.9,by=0.1)
xqs <- quantile(simdf$x, probs = (1:(L-1))/L)
names(xqs) <- c()
qqp  <- matrix(0, nrow=length(xqs), ncol=length(qns))
i <- 1
for (qn in qns){
    qremFit <-  QREM(lm,linmod=y~x*x2 +x3, df=simdf, qn=qn)
    qrdg <- QRdiagnostics(simdf$x, "x",qremFit$ui, qn,  plot.it = FALSE)
    for (j in 1:(L-1)) {
        qqp[j,i] <- length(which(qrdg$y < xqs[j])) / length(which(qrdg$x < xqs[j]))
    }
    i <- i+1
}
flatQQplot(simdf$x,xqs,qqp,qns)
}
}
