\name{simdf}
\alias{simdf}
\docType{data}
\title{
simdf
}
\description{
A simulated data set for the QREM package.
}
\usage{data("simdf")}
\format{
  A data frame with 10,000 simulated observations with the following 4 variables.
  \describe{
    \item{\code{y}}{a numeric response}
    \item{\code{x}}{first numeric predictor}
    \item{\code{x2}}{second numeric predictor}
    \item{\code{x3}}{a three-level categorical predictor}
  }
}
\examples{
data(simdf)
}
\keyword{datasets}
