% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QREM.R
\name{bcov}
\alias{bcov}
\title{Covariance matrix estimation}
\usage{
bcov(qremFit, linmod, dframe, qn, userwgts = NULL)
}
\arguments{
\item{qremFit}{A fitted model object, returned by QREM.}

\item{linmod}{A formula (the linear model for fitting in the M step).}

\item{dframe}{A data frame containing the columns in the formula.}

\item{qn}{The selected quantile. Must be in (0,1).}

\item{userwgts}{The user-provided sampling weights (optional. Default=NULL.)}
}
\value{
A covariance matrix for the fixed-effects model.
}
\description{
Covariance matrix estimation for the fixed effects case using Bahadur's
representation. In mixed model, we use the BLUPs (as if they were estimated
as fixed effects) and the same formula is used.
}
\examples{
data(simdf)
qremFit <-  QREM(lm,linmod=y~x*x2 +x3, df=simdf, qn=0.2)
covmat <- bcov(qremFit,linmod=y~x*x2 +x3, df=simdf, qn=0.2)
}
