% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QREM.R
\name{flatQQplot}
\alias{flatQQplot}
\title{A `flat QQ-plot' for a fitted quantile regression model.}
\usage{
flatQQplot(
  dat,
  cnum = NULL,
  vname = NULL,
  qrfits,
  qns,
  maxm = 30,
  sdevs = 4,
  filename = NULL,
  plot.it = TRUE
)
}
\arguments{
\item{dat}{The data matrix.}

\item{cnum}{The selected column number.}

\item{vname}{The selected variable name (if cnum is not specified).}

\item{qrfits}{A list of QR-fitted models.}

\item{qns}{The quantiles.}

\item{maxm}{The maximum number of segments in the partision of the selected variable (default=30)}

\item{sdevs}{Used to determine the ranges for critical values under the null model and thus the range of colors in the heatmap (default= 4).}

\item{filename}{The pdf file to save the plot. Default is NULL (print to the screen.)}

\item{plot.it}{A logical variable used to determine whether to show the diagnostic plot (TRUE).}
}
\description{
Showing multiple ('flat') QQ plots for different quantiles using a heatmap.
}
\examples{
\donttest{
data(simdf)
qns <- seq(0.1,0.9,by=0.1)
qrfits <- list()
for (i in 1:length(qns)) {
 qrfits[[i]] <- QREM(lm,linmod=y~x*x2 +x3, df=simdf, qn=qns[i])
}
pvals <- flatQQplot(dat=simdf, cnum = 2, qrfits=qrfits, qns=qns, maxm = 20, plot.it = TRUE)
pvals <- flatQQplot(dat=simdf, cnum = 3, qrfits=qrfits, qns=qns, maxm = 20, plot.it = TRUE)
pvals <- flatQQplot(dat=simdf, cnum = 4, qrfits=qrfits, qns=qns, maxm = 20, plot.it = TRUE)
}
}
