% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QREM.R
\name{QRdiagnostics}
\alias{QRdiagnostics}
\title{Quantile regression diagnostics based on the residuals, u_i.}
\usage{
QRdiagnostics(X, varname, u_i, qn, plot.it = TRUE, filename = NULL)
}
\arguments{
\item{X}{A predictor included in the regression.}

\item{varname}{The predictor's name (for plotting).}

\item{u_i}{The QR residuals.}

\item{qn}{The quantile used in the regression.}

\item{plot.it}{Boolean, if TRUE, will show the histogram of the residuals and the fitted kernel density estimate (default=TRUE).}

\item{filename}{The pdf file to save the plot. Default is NULL (print to the screen.)}
}
\value{
A list, as follows, plus the marginal deviance:
\itemize{
  \item For a continuous predictor, the list is called qqp and contains the output from qqplot().
  \item For a categorical variable, the list is called qqlvl, and it contains the empirical percentages of points below the regression line, for each level.
}
}
\description{
For a given predictor, create a qq-plot if continuous, or a spinogram if categorical.
}
\examples{
data(simdf)
qremFit <-  QREM(lm,linmod=y~x*x2 +x3, df=simdf, qn=0.2)
qrdg <- QRdiagnostics(simdf$x, "x",qremFit$ui, 0.2,  plot.it = TRUE)
qrdg <- QRdiagnostics(simdf$x3, "x3",qremFit$ui, 0.2,  plot.it = TRUE)
}
