% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QREM.R
\name{QREM}
\alias{QREM}
\title{Fitting a quantile regression model via the EM algorithm.}
\usage{
QREM(
  func,
  linmod,
  dframe,
  qn,
  userwgts = NULL,
  ...,
  err = 10,
  maxit = 1000,
  tol = 0.001,
  maxInvLambda = 300
)
}
\arguments{
\item{func}{The fitting function (lm, lmer, or gam).}

\item{linmod}{A formula (the linear model for fitting in the M step).}

\item{dframe}{A data frame containing the columns in the formula.}

\item{qn}{The selected quantile. Must be in (0,1).}

\item{userwgts}{The user-provided sampling weights (optional. Default=NULL.)}

\item{...}{Any arguments to be passed to func (except for the formula and weights). Note that gam requires the family of the error distribution.}

\item{err}{The initial value for the estimation error (default=10). Must be greater than tol (below).}

\item{maxit}{The maximum number of EM iterations (default=1000).}

\item{tol}{The error tolerance level (default=0.001).}

\item{maxInvLambda}{The maximum value of the weight for WLS fitting (default=300).}
}
\value{
A list containing the following
\itemize{
  \item coef The estimated regression coefficients (a list).
  \item fitted.mod The output from lm(), lmer(), or gam() in the last EM iteration.
  \item empq The percentage of points below the regression line (the empirical quantile).
  \item ui The quantile regression residuals.
  \item weights The weights used in the WLS solution.
  \item iter The number of EM iterations.
  \item err The final EM error (convergence criterion).
}
}
\description{
Fitting a quantile regression model via the EM algorithm.
}
\examples{
data(simdf)
qremFit <-  QREM(lm,linmod=y~x*x2 +x3, df=simdf, qn=0.2)
summary(aov(qremFit$fitted.mod))
summary(qremFit$fitted.mod)$coef
}
